3-e2e-drug-discovery-sample/backend/requirements.txt
# Azure AI Foundry SDKs
azure-ai-projects>=1.0.0b5
azure-ai-inference[opentelemetry]>=1.0.0b5
azure-ai-evaluation>=1.0.0b5
azure-search-documents==11.4.0
azure-identity==1.16.0

# FastAPI and Web Server
fastapi>=0.109.2,<0.110.0
uvicorn>=0.27.1,<0.28.0
python-dotenv>=1.0.0,<2.0.0

# OpenTelemetry Integration
opentelemetry-api>=1.29.0,<2.0.0
opentelemetry-sdk>=1.29.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.44b0,<1.0.0
opentelemetry-instrumentation>=0.44b0,<1.0.0
azure-core-tracing-opentelemetry>=1.0.0b11,<2.0.0
opentelemetry-exporter-otlp>=1.29.0,<2.0.0
opentelemetry-semantic-conventions>=0.50b0,<1.0.0

# Models
pydantic==2.5.2

# Data Analysis
pandas>=2.1.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Testing
pytest==8.3.4
pytest-mock==3.14.0
pytest-asyncio==0.25.2
httpx==0.27.0

# Utilities
requests==2.31.0
python-jose[cryptography]==3.3.0  # For JWT handling
